requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.28.38
botocore==1.31.38
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-cors-headers==4.2.0
django-extensions==3.2.3
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
dynaconf==3.2.2
hvac==1.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.3
msgpack==1.0.5
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyhcl==0.4.4
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
s3transfer==0.6.2
segno==1.5.2
service-identity==21.1.0
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.16
validators==0.21.2
whitenoise==6.5.0
zope.interface==6.0
